@@ -1,120 +1,144 @@
 {
-  "name": "pie-person",
-  "version": "1.0.0",
-  "description": "",
-  "main": "/server/server.js",
-  "scripts": {
-    "indexer": "nodemon --exec \"ts-node --project ./indexer/tsconfig.json ./indexer\" --trace-warnings --watch ./indexer --legacy-watch",
-    "watch": "webpack-dev-server --color --progress -w --config ./webpack/webpack.config.js",
-    "build": "webpack --color --progress --config ./webpack/webpack.config.js",
-    "build:production": "webpack --color --progress -p --config ./webpack/webpack.config.js",
-    "start": "nodemon --exec ts-node ./server/bin/www --trace-warnings --watch ./server --ignore ./app --ignore ./server",
-    "test": "jest --runInBand --cache",
-    "test:watch": "npm test -- --watch"
-  },
+  "name": "pieperson",
+  "description": "Path of Exile API App",
+  "version": "0.0.1",
+  "private": true,
+  "license": "GPL-3.0",
+  "author": "Will Schwanke",
   "repository": {
     "type": "git",
-    "url": "local"
+    "url": "https://bitbucket.org/wschwanke/pieperson.git"
   },
-  "keywords": [
-    "pie"
-  ],
-  "author": "Will Schwanke",
-  "license": "MIT",
-  "nodemonConfig": {
-    "verbose": "true"
+  "engines": {
+    "node": "10.x"
+  },
+  "scripts": {
+    "clear": "rm -rf ./node_modules/",
+    "clear:build": "rm -rf ./.build",
+    "clear:cache": "rm -rf ./node_modules/.cache",
+    "api:start": "ts-node --typeCheck --project ./tsconfig.api.json ./api/server.ts",
+    "api:build": "tsc --build ./tsconfig.api.json",
+    "api:dev": "ts-node-dev -r tsconfig-paths/register --respawn --project ./tsconfig.api.json ./api/keto-tracker-api.ts --typeCheck",
+    "client:build": "webpack --color --progress -p --config ./webpack/webpack.config.js",
+    "client:dev": "yarn run clear:cache && webpack-dev-server --color --progress -w --config ./webpack/webpack.config.js",
+    "webserver:start": "ts-node --typeCheck --project ./tsconfig.webserver.json ./webserver/server.ts",
+    "webserver:build": "tsc --build ./tsconfig.webserver.json",
+    "webserver:dev": "ts-node-dev -r tsconfig-paths/register --respawn --project ./tsconfig.webserver.json ./webserver/keto-tracker-webserver.ts --watch ./webserver --typeCheck",
+    "build": "yarn run webserver:build && yarn run client:build",
+    "start": "node ./.build/webserver/server.js",
+    "ngrok": "ngrok http --subdomain wschwanke 3001",
+    "ngrok:test": "ngrok http --subdomain wschwanke 3000"
   },
   "dependencies": {
-    "@types/axios": "^0.14.0",
-    "@types/dotenv": "^6.1.0",
-    "@types/isomorphic-fetch": "^0.0.34",
-    "@types/mongodb": "^3.1.15",
-    "@types/winston": "^2.4.4",
-    "axios": "^0.17.1",
-    "body-parser": "^1.17.2",
-    "classnames": "^2.2.6",
-    "cookie-parser": "^1.4.3",
-    "dotenv": "^4.0.0",
-    "express": "^4.16.2",
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.2.2",
+    "@babel/plugin-proposal-class-properties": "^7.3.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.3.2",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.2.0",
+    "@babel/plugin-transform-typescript": "^7.3.2",
+    "@babel/polyfill": "^7.2.5",
+    "@babel/preset-env": "^7.3.1",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-typescript": "^7.1.0",
+    "babel-loader": "^8.0.4",
+    "bootstrap": "^4.3.1",
+    "classnames": "^2.2.5",
+    "copy-webpack-plugin": "^5.0.3",
+    "dotenv": "^6.2.0",
+    "graphql": "^14.3.0",
+    "ioredis": "^4.10.0",
     "isomorphic-fetch": "^2.2.1",
+    "jquery": "^3.4.1",
     "js-cookie": "^2.2.0",
-    "jsdom": "^13.0.0",
-    "lodash": "^4.17.11",
-    "minimist": "^1.2.0",
-    "moment": "^2.22.1",
-    "mongodb": "^3.0.3",
-    "morgan": "^1.9.0",
-    "nodemon": "^1.11.0",
-    "path": "^0.12.7",
-    "react": "^16.7.0-alpha.0",
-    "react-dom": "^16.7.0-alpha.0",
-    "react-helmet": "^5.2.0",
-    "react-redux": "^5.1.0",
-    "react-router-dom": "^4.3.1",
-    "redux": "^4.0.1",
-    "redux-thunk": "^2.3.0",
-    "simplebar": "^3.1.0-beta.5",
-    "tslint": "^5.11.0",
+    "jsdom": "^11.12.0",
+    "jsonwebtoken": "^8.5.1",
+    "koa": "^2.6.2",
+    "koa-bodyparser": "^4.2.1",
+    "koa-helmet": "^4.2.0",
+    "koa-logger": "^3.2.0",
+    "koa-mount": "^4.0.0",
+    "koa-router": "^7.4.0",
+    "koa-static": "^5.0.0",
+    "lodash": "^4.17.4",
+    "node-fetch": "^2.4.1",
+    "popper.js": "^1.15.0",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-hot-loader": "^4.6.5",
+    "react-loadable": "^5.5.0",
+    "react-router-dom": "^5.0.1",
+    "react-test-renderer": "^16.4.2",
+    "ts-node": "^8.3.0",
+    "typescript": "^3.5.2",
     "whatwg-fetch": "^3.0.0",
-    "winston": "^3.1.0"
+    "winston": "^3.2.1"
   },
   "devDependencies": {
-    "@babel/cli": "^7.1.2",
-    "@babel/core": "^7.1.2",
-    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
-    "@babel/polyfill": "^7.0.0",
-    "@babel/preset-env": "^7.1.0",
-    "@babel/preset-react": "^7.0.0",
+    "@hot-loader/react-dom": "^16.8.6",
     "@types/classnames": "^2.2.6",
+    "@types/dotenv": "^6.1.0",
     "@types/enzyme": "^3.1.13",
     "@types/enzyme-adapter-react-16": "^1.0.3",
     "@types/fetch-mock": "^6.0.3",
+    "@types/ioredis": "^4.0.12",
+    "@types/isomorphic-fetch": "^0.0.35",
+    "@types/jest": "^23.3.1",
+    "@types/js-cookie": "^2.2.0",
     "@types/jsdom": "^11.0.6",
+    "@types/koa": "^2.0.47",
+    "@types/koa-bodyparser": "^4.2.1",
+    "@types/koa-helmet": "^3.1.2",
+    "@types/koa-logger": "^3.1.1",
+    "@types/koa-mount": "^3.0.1",
+    "@types/koa-router": "^7.0.35",
+    "@types/koa-session": "^5.7.4",
+    "@types/koa-static": "^4.0.0",
     "@types/lodash": "^4.14.116",
-    "@types/node": "^10.12.5",
+    "@types/node": "^10.12.9",
+    "@types/node-fetch": "^2.3.3",
     "@types/react": "^16.4.11",
     "@types/react-dom": "^16.0.7",
-    "@types/react-helmet": "^5.0.7",
-    "@types/react-loadable": "^5.4.1",
-    "@types/react-redux": "^6.0.9",
+    "@types/react-hot-loader": "^4.1.0",
+    "@types/react-loadable": "^5.5.1",
     "@types/react-router-dom": "^4.3.0",
-    "@types/redux-logger": "^3.0.6",
-    "@types/redux-mock-store": "^1.0.0",
-    "autoprefixer": "^9.3.1",
-    "babel-loader": "^8.0.4",
-    "babel-preset-env": "^1.6.1",
-    "cache-loader": "^1.2.5",
-    "clean-webpack-plugin": "^0.1.19",
-    "copy-webpack-plugin": "^4.6.0",
-    "css-loader": "^1.0.1",
-    "cssnano": "^4.1.7",
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.7.0",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "fork-ts-checker-webpack-plugin": "^0.4.15",
+    "@types/webpack": "^4.4.24",
+    "@typescript-eslint/parser": "^1.3.0",
+    "autoprefixer": "^9.4.7",
+    "clean-webpack-plugin": "^1.0.1",
+    "css-loader": "^2.1.0",
+    "cssnano": "^4.1.8",
+    "enzyme": "^3.6.0",
+    "enzyme-adapter-react-16": "^1.2.0",
+    "eslint": "^4.19.1 || ^5.3.0",
+    "eslint-plugin-react-hooks": "^1.0.1",
+    "extract-css-chunks-webpack-plugin": "^3.3.2",
+    "file-loader": "^4.0.0",
+    "fork-ts-checker-webpack-plugin": "^0.5.2",
     "friendly-errors-webpack-plugin": "^1.7.0",
-    "happypack": "^5.0.0",
-    "hard-source-webpack-plugin": "^0.12.0",
+    "hard-source-webpack-plugin": "^0.13.1",
     "html-webpack-harddisk-plugin": "^1.0.1",
     "html-webpack-plugin": "^3.2.0",
-    "jest": "^23.6.0",
-    "mini-css-extract-plugin": "^0.4.4",
-    "node-sass": "^4.10.0",
-    "postcss-loader": "^2.0.6",
-    "postcss-safe-parser": "^3.0.1",
-    "react-hot-loader": "^4.3.12",
-    "redux-devtools": "^3.4.1",
-    "redux-logger": "^3.0.6",
+    "jest": "^23.5.0",
+    "mini-css-extract-plugin": "^0.5.0",
+    "node-sass": "^4.11.0",
+    "nodemon": "^1.18.6",
+    "postcss-loader": "^3.0.0",
+    "postcss-safe-parser": "^4.0.1",
     "sass-loader": "^7.1.0",
     "style-loader": "^0.23.1",
-    "ts-loader": "^5.3.0",
-    "ts-node": "^7.0.1",
-    "typescript": "^3.1.6",
-    "webpack": "^4.25.1",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.4",
-    "webpack-merge": "^4.1.4",
-    "write-file-webpack-plugin": "^4.4.1"
-  }
+    "supertest": "^4.0.2",
+    "ts-jest": "^23.1.3",
+    "ts-node-dev": "^1.0.0-pre.40",
+    "tsconfig-paths": "^3.8.0",
+    "tslint": "^5.11.0",
+    "tslint-react": "^3.6.0",
+    "typescript-tslint-plugin": "^0.3.1",
+    "webpack": "^4.29.3",
+    "webpack-cli": "^3.2.3",
+    "webpack-dev-server": "^3.1.14",
+    "webpack-merge": "^4.2.1",
+    "write-file-webpack-plugin": "^4.5.0"
+  },
+  "peerDependencies": {}
 }
