@@ -1,42 +1,17 @@
-/** @format */
-
-const webpack = require('webpack');
 const { join } = require('path');
-const HappyPack = require('happypack');
-const merge = require('webpack-merge');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
 const FriendlyErrorsWebpackPlugin = require('friendly-errors-webpack-plugin');
-const ForkTsCheckerPlugin = require('fork-ts-checker-webpack-plugin');
-const HardSourceWebpackPlugin = require('hard-source-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const HtmlWebpackHarddiskPlugin = require('html-webpack-harddisk-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const ExtractCssChunks = require('extract-css-chunks-webpack-plugin');
+const ForkTsCheckerPlugin = require('fork-ts-checker-webpack-plugin');
+const HardSourceWebpackPlugin = require('hard-source-webpack-plugin');
+const webpack = require('webpack');
+const merge = require('webpack-merge');
+
+const memoryLimit = 2048;
 
 const config = require('./config');
-const happyThreadPool = HappyPack.ThreadPool({ size: 3 });
-
-/**
- * This checks to see if we are NOT watching with webpack and returns an array where MiniCssExtractPlugin
- * is loaded before our happypack loader so it doesn't cause an error.
- */
-function isWatchingHappyStylesRules(happyId) {
-  if (!config.enabled.watcher) {
-    return [MiniCssExtractPlugin.loader, 'happypack/loader?id='.concat(happyId)]
-  }
-  return 'happypack/loader?id='.concat(happyId);
-}
-
-/**
- * This checks to see if we are watching with webpack and then unshifts style-loader to the front of our loaders
- * so css is injected into our browser.
- */
-function isWatchingHappyStylesPlugins(loaders) {
-  if (config.enabled.watcher) {
-    return [{ loader: 'style-loader' }].concat(loaders);
-  }
-  return loaders;
-}
 
 let webpackConfig = {
   context: config.paths.root,
@@ -45,7 +20,11 @@ let webpackConfig = {
 
   entry: {
     bundle: [
-      join(config.paths.root, 'client/index.tsx'),
+      '@babel/polyfill',
+      join(config.paths.root, 'client/entry.tsx'),
+      join(config.paths.root, 'client/assets/font-awesome/fontawesome.scss'),
+      join(config.paths.root, 'client/assets/font-awesome/regular.scss'),
+      join(config.paths.root, 'client/assets/font-awesome/solid.scss'),
       join(config.paths.root, 'client/assets/styles/main.scss'),
     ],
   },
@@ -59,7 +38,7 @@ let webpackConfig = {
   },
 
   output: {
-    path: config.paths.dist,
+    path: config.paths.compiled,
     pathinfo: false,
     publicPath: config.paths.public,
     filename: !config.env.production ? 'scripts/[name].js' : 'scripts/[name].[hash].js',
@@ -69,38 +48,102 @@ let webpackConfig = {
   module: {
     rules: [
       {
-        test: /\.ts$/,
-        exclude: [/bower_components/, /node_modules/],
-        use: 'happypack/loader?id=ts'
-      },
-      {
-        test: /\.tsx$/,
-        exclude: [/bower_components/, /node_modules/],
-        use: 'happypack/loader?id=tsx'
+        test: /\.t(s|sx)$/,
+        exclude: [/(node_modules)/, join(config.paths.root, 'server'), join(config.paths.root, 'indexer')],
+        use: [
+          {
+            loader: 'babel-loader'
+          },
+        ],
       },
       {
         test: /\.(sc|sa)ss$/,
-        exclude: /node_modules/,
-        use: isWatchingHappyStylesRules('scss')
+        use: [
+          {
+            loader: ExtractCssChunks.loader,
+            options: {
+              hot: config.enabled.watcher,
+              reloadAll: true,
+            },
+          },
+          {
+            loader: 'css-loader',
+            options: {
+              sourceMap: config.enabled.sourceMaps
+            }
+          },
+          {
+            loader: 'postcss-loader',
+            options: {
+              config: {
+                path: join(config.paths.root, '/webpack'),
+                ctx: config
+              },
+              sourceMap: config.enabled.sourceMaps,
+            },
+          },
+          {
+            loader: 'sass-loader',
+            options: {
+              includePaths: [
+                join(config.paths.root, '/node_modules'),
+              ],
+              sourceMap: config.enabled.sourceMaps,
+            },
+          },
+        ]
       },
       {
         test: /\.css$/,
-        exclude: /node_modules/,
-        use: isWatchingHappyStylesRules('css')
+        use: [
+          {
+            loader: 'style-loader',
+          },
+          { loader: 'css-loader', options: { sourceMap: config.enabled.sourceMaps } },
+          {
+            loader: 'postcss-loader',
+            options: {
+              config: { path: join(config.paths.root, '/webpack'), ctx: config },
+              sourceMap: config.enabled.sourceMaps,
+            },
+          },
+        ]
+      },
+      {
+        test: /\.(png|jpe?g|gif|svg)$/,
+        use: [
+          {
+            loader: 'file-loader',
+            options: {
+              name: 'images/[name].[ext]',
+            },
+          },
+        ],
+      },
+      {
+        test: /\.(eot|woff|ttf|svg|woff2)$/,
+        use: [
+          {
+            loader: 'file-loader',
+            options: {
+              name: 'fonts/[name].[ext]',
+            },
+          },
+        ],
       },
     ],
   },
 
   resolve: {
     alias: {
-      blocks: join(config.paths.root, 'client/blocks'),
-      components: join(config.paths.root, 'client/components'),
-      layout: join(config.paths.root, 'client/layout'),
-      lib: join(config.paths.root, 'client/lib'),
-      scenes: join(config.paths.root, 'client/scenes'),
-      router: join(config.paths.root, 'client/router'),
-      state: join(config.paths.root, 'client/state'),
-      test: join(config.paths.root, 'client/test'),
+      "@Assets": join(config.paths.root, 'client/assets'),
+      "@Blocks": join(config.paths.root, 'client/blocks'),
+      "@Components": join(config.paths.root, 'client/components'),
+      "@Layout": join(config.paths.root, 'client/layout'),
+      "@Lib": join(config.paths.root, 'client/lib'),
+      "@Scenes": join(config.paths.root, 'client/scenes'),
+      "@Router": join(config.paths.root, 'client/router'),
+      "@Test": join(config.paths.root, 'client/test'),
     },
     extensions: ['.js', '.jsx', '.ts', '.tsx', '.css', '.scss'],
   },
@@ -109,9 +152,9 @@ let webpackConfig = {
     splitChunks: {
       chunks: 'all',
       minSize: 30000,
-      minChunks: 1,
-      maxAsyncRequests: 5,
-      maxInitialRequests: 3,
+      minChunks: 2,
+      maxAsyncRequests: 6,
+      maxInitialRequests: 5,
       automaticNameDelimiter: '-',
       name: true,
       cacheGroups: {
@@ -131,132 +174,42 @@ let webpackConfig = {
   target: 'web',
 
   plugins: [
-    // Build caching plugin for webpack
+    // Removes files from public/dist directory before build
     new HardSourceWebpackPlugin(),
+    new webpack.ProgressPlugin(),
     new ForkTsCheckerPlugin({
-      memoryLimit: 512,
+      tsconfig: join(config.paths.root, 'tsconfig.client.json'),
+      memoryLimit: memoryLimit,
       checkSyntacticErrors: true,
       async: false,
-      tsconfig: join(config.paths.root, 'client'),
       watch: [join(config.paths.root, 'client')],
       workers: ForkTsCheckerPlugin.ONE_CPU,
     }),
-    // Removes files from public/dist directory before build
-    new CleanWebpackPlugin([config.paths.dist], {
+    new CleanWebpackPlugin([config.paths.clean], {
       root: config.paths.root,
       verbose: false,
     }),
-    new HappyPack({
-      id: 'ts',
-      threadPool: happyThreadPool,
-      loaders: [
-        {
-          loader: 'babel-loader',
-          options: {
-            cacheDirectory: false,
-            babelrc: false,
-            plugins: ['@babel/plugin-syntax-dynamic-import', 'react-hot-loader/babel'],
-          },
-        },
-        {
-          loader: 'ts-loader',
-          options: {
-            experimentalFileCaching: true,
-            transpileOnly: true,
-            happyPackMode: true,
-          },
-        },
-      ],
-    }),
-    new HappyPack({
-      id: 'tsx',
-      threadPool: happyThreadPool,
-      loaders: [
-        {
-          loader: 'babel-loader',
-          options: {
-            cacheDirectory: false,
-            babelrc: false,
-            plugins: ['@babel/plugin-syntax-dynamic-import', 'react-hot-loader/babel'],
-          },
-        },
-        {
-          loader: 'ts-loader',
-          options: {
-            experimentalFileCaching: true,
-            transpileOnly: true,
-            happyPackMode: true,
-          },
-        },
-      ],
-    }),
-    new HappyPack({
-      id: 'css',
-      threadPool: happyThreadPool,
-      loaders: isWatchingHappyStylesPlugins([
-        { loader: 'css-loader', options: { sourceMap: config.enabled.sourceMaps } },
-        {
-          loader: 'postcss-loader',
-          options: {
-            config: { path: join(config.paths.root, '/webpack'), ctx: config },
-            sourceMap: config.enabled.sourceMaps,
-          },
-        },
-      ]),
-    }),
-    new HappyPack({
-      id: 'scss',
-      threadPool: happyThreadPool,
-      loaders: isWatchingHappyStylesPlugins([
-        { loader: 'css-loader', options: { sourceMap: config.enabled.sourceMaps } },
-        {
-          loader: 'postcss-loader',
-          options: {
-            config: { path: join(config.paths.root, '/webpack'), ctx: config },
-            sourceMap: config.enabled.sourceMaps,
-          },
-        },
-        {
-          loader: 'sass-loader',
-          options: {
-            includePaths: [
-              join(config.paths.root, '/node_modules'),
-            ],
-            sourceMap: config.enabled.sourceMaps,
-          },
-        },
-      ]),
-    }),
-    // Copies all the images from the assets folder and moves them to the public folder
-    new CopyWebpackPlugin(
-      [
-        {
-          context: join(config.paths.assets, 'images'),
-          from: '**/*',
-          to: join(config.paths.root, 'public/dist/images'),
-          flatten: true,
-        },
-        {
-          context: join(config.paths.assets, 'fonts'),
-          from: '**/*',
-          to: join(config.paths.root, 'public/dist/fonts'),
-          flatten: true,
-        },
-      ],
+    // new CopyWebpackPlugin([
+    //     { from: join(config.paths.root, 'client/assets/fonts'), to: join(config.paths.root, '.build/client/assets/fonts')},
+    //     { from: join(config.paths.root, 'client/assets/images'), to: join(config.paths.root, '.build/client/assets/images')},
+    //   ]
+    // ),
+    new ExtractCssChunks(
       {
-        copyUnmodified: true,
-      },
+        // Options similar to the same options in webpackOptions.output
+        // both options are optional
+        filename: !config.env.production ? 'styles/[name].css' : 'styles/[name].[hash].css',
+        chunkFilename: !config.env.production ? 'styles/[name].css' : 'styles/[name].[hash].css',
+        hot: true, // if you want HMR - we try to automatically inject hot reloading but if it's not working, add it to the config
+        orderWarning: false, // Disable to remove warnings about conflicting order between imports
+        reloadAll: true, // when desperation kicks in - this is a brute force HMR flag
+      }
     ),
-    // Extracts the css from webpack and saves it as a file in public/dist
-    new MiniCssExtractPlugin({
-      filename: !config.env.production ? 'styles/[name].css' : 'styles/[name].[hash].css',
-      chunkFilename: !config.env.production ? 'styles/[name].css' : 'styles/[name].[hash].css',
-    }),
     // Generates an html file that has all the required assets built by webpack injected into it
     new HtmlWebpackPlugin({
       inject: false,
-      filename: '../_index.html',
-      template: join(config.paths.root, 'client/index.html'),
+      filename: '../index.html',
+      template: join(config.paths.root, 'client/layout/index.html'),
       alwaysWriteToDisk: true,
     }),
     // Creates physical files on the hard disk during webpack-dev-server watch
@@ -267,16 +220,19 @@ let webpackConfig = {
   cache: false,
 };
 
-if (config.enabled.linting) {
-  webpackConfig.module.rules.push(require('./webpack.linter.config.js'));
-}
-
 if (config.env.production) {
   webpackConfig.plugins.push(new webpack.NoEmitOnErrorsPlugin());
 }
 
 if (config.enabled.watcher) {
   webpackConfig = merge(webpackConfig, require('./webpack.watcher.config.js'));
+  webpackConfig = merge(webpackConfig, {
+    resolve: {
+      alias: {
+        'react-dom': '@hot-loader/react-dom',
+      },
+    },
+  });
 }
 
 module.exports = webpackConfig;
